imaprt_api/requirements.txt
asgiref==3.2.9
certifi==2020.4.5.2
chardet==3.0.4
coverage==5.1
Django==3.0.7
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
numpy==1.18.5
opencv-python==3.4.9.33
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
wincertstore==0.2
